+++ changelog.md
@@ -1,6 +1,7 @@
 # changelog
 
 ## 1.16.2
+* `NEW` diagnostic: `duplicate-set-field`
 * `FIX` [#406](https://github.com/sumneko/lua-language-server/issues/406)
 
 ## 1.16.1